+++ codemp/game/NPC_AI_ImperialProbe.c
@@ -82,9 +82,9 @@ void ImperialProbe_MaintainHeight( void )
 		dif = NPCS.NPC->enemy->r.currentOrigin[2] - NPCS.NPC->r.currentOrigin[2];
 
 		// cap to prevent dramatic height shifts
-		if ( fabs( dif ) > 8 )
+		if ( fabsf( dif ) > 8 )
 		{
-			if ( fabs( dif ) > 16 )
+			if ( fabsf( dif ) > 16 )
 				dif = ( dif < 0 ? -16 : 16 );
 
 			NPCS.NPC->client->ps.velocity[2] = (NPCS.NPC->client->ps.velocity[2]+dif)/2;
@@ -103,7 +103,7 @@ void ImperialProbe_MaintainHeight( void )
 		{
 			dif = goal->r.currentOrigin[2] - NPCS.NPC->r.currentOrigin[2];
 
-			if ( fabs( dif ) > 24 ) {
+			if ( fabsf( dif ) > 24 ) {
 				NPCS.ucmd.upmove = ( NPCS.ucmd.upmove < 0 ? -4 : 4 );
 			}
 			else {
@@ -111,7 +111,7 @@ void ImperialProbe_MaintainHeight( void )
 				{
 					NPCS.NPC->client->ps.velocity[2] *= VELOCITY_DECAY;
 
-					if ( fabs( NPCS.NPC->client->ps.velocity[2] ) < 2 )
+					if ( fabsf( NPCS.NPC->client->ps.velocity[2] ) < 2 )
 						NPCS.NPC->client->ps.velocity[2] = 0;
 				}
 			}
@@ -161,7 +161,7 @@ void ImperialProbe_MaintainHeight( void )
 	{
 		NPCS.NPC->client->ps.velocity[0] *= VELOCITY_DECAY;
 
-		if ( fabs( NPCS.NPC->client->ps.velocity[0] ) < 1 )
+		if ( fabsf( NPCS.NPC->client->ps.velocity[0] ) < 1 )
 		{
 			NPCS.NPC->client->ps.velocity[0] = 0;
 		}
@@ -171,7 +171,7 @@ void ImperialProbe_MaintainHeight( void )
 	{
 		NPCS.NPC->client->ps.velocity[1] *= VELOCITY_DECAY;
 
-		if ( fabs( NPCS.NPC->client->ps.velocity[1] ) < 1 )
+		if ( fabsf( NPCS.NPC->client->ps.velocity[1] ) < 1 )
 		{
 			NPCS.NPC->client->ps.velocity[1] = 0;
 		}
